adicionando ElasticEntity para futuro reflexion - Java

+++ xtarefas/src/main/java/com/basis/campina/xtarefas/domain/elastic/BaseDocument.java
@@ -21,7 +21,7 @@ public class BaseDocument implements Serializable {
     protected static final String SORT = "sort";
     protected static final String DATE_PATTERN = "dd/MM/yyyy";
 
-    @MultiField(mainField = @Field(type = FieldType.Long, store = true),
+    @MultiField(mainField = @Field(type = FieldType.Integer, store = true),
             otherFields = {@InnerField(suffix = SORT, type = FieldType.Integer, store = true)})
     protected Integer id;
